Backwards incompatible change, see full commit message for details.
Somehow I missed the point, where there were input and output errors. It was
more or less handled in the C version, but all of the errors are totally ignored
in the Go version. This commit adds error. The only incompatible change is
PollEvent function, you'll have to handle or ignore its second return value.

I think an ability to handle these errors is important, not very important when
running a termbox application locally, but it is important when running through
ssh on a remote host. Hopefully now it's the very last incompatible API change.