W32: Don't use a separate console screen. Stop event producer on Close().
It seems that this is what most major console programs do. Yes it's a bit
destructive, because it clears all the contents of the console buffer. But at
the same time it behaves nicely in all cases. Like running under ConsoleZ
(tabbed console app for windows).

This fix also tries to make a proper shutdown possible on windows. It creates
a special event object and uses WaitForMultipleObjects instead of just
blocking on console input. This way it's possible to kill the event producing
goroutine gracefully. Not sure how it'll end up working in practice, but let's
give it a try.

Fixes #23.
Fixes #58.