Improve portability for POSIX systems
We use simple CGO style functions for the low level accesses to
set or get termios attributes and the TIOCGWINSIZE ioctl.  This
helps for systems where SYS_ioctl may be absent (such as solaris),
while stil keeping POSIX compatibility.  (Note however that the
TIOCGWINSIZE call may be missing on some systems. In such a case
we provide a default guess of 80x24, unless environment variables
COLUMNS and LINES indicate otherwise.)

Additionally, there is no need for non-blocking I/O.  With Go,
we can use os.Read() in a goroutine, and it will return as soon
as data is available, even if the buffer is not open.  This helps
with both portability (O_ASYNC and F_SETOWN aren't universally
available), and readability.

Btw, these changes also allow us to eliminate the use of the
unsafe pointers in the Go code.

I've tested the included demo programs, as well as my own program
on both Darwin and illumos/solaris systems.  Furthermore, I've
tested the topsl (github.com/gdamore/topsl) demos with these
changes on both systems.

Note that on solaris this requires the use of go v1.5 or better,
as cgo is not present on older versions of that port.  All POSIX
platforms supported by Go 1.5 also support cgo.