Fix support for variable width UTF-8 characters
It turns out that some characters in Asian locales (CJK)
can be double wide, occupying two cells on normal terminals.
This is necessary as such characters are too complex to draw
in a typical 8x8 or even 8x16 pixels cell.  Additionally some
characters are combining marks, which have to be used together
with another character, to form the glyph that is seen on screen.
(That is, there can be multiple runes that combine to create a
single onscreen glyph.)

This change fixes the demo, to support wide characters properly.

It also changes the demo to reject zero width (combining marks)
characters.  This is necessary, since otherwise we could have
a situation where the the backing array is not sufficiently large
to hold all the data on screen.  To support combining marks, such
as \U+02D8, we would need to update termbox to keep track of
combining marks, and somehow ensure that we didn't append more
such marks than the backing store allows.  As such marks are not
necessary in any major language (where they arise there are
precomposed forms that are preferable), we simply decline to
accept them here.

In the future, a superior way to handle these combining marks might
be to automatically attempt to compose with the charcter under the
rune, to create a new precomposed form.  The reality is that this
is probably also not worth while.

This fixes #59.