Fix crashes and odd behaviour due to builtin terminfo inconsistencies
In recent commits

- a08d44e997 implemented dim, blink hidden
- b53f34a171 extra attributes wip

3 more attributes were added - hidden, dim and cursive.

Unfortunately these commits didn't adjust the builtin terminfo.

This leads to crashes which look like this on quit

    panic: runtime error: index out of range [14] with length 14
    goroutine 1 [running]:
    github.com/nsf/termbox-go.Close()
        github.com/nsf/termbox-go@v1.1.0/api.go:146 +0x458

And other odd behaviour like arrow keys not being recognised.

This patch fixes the terminfo, and names all the entries to make it
easier to update. It also adds a test to make sure the terminfo
doesn't get out of sync again.

See: https://github.com/rclone/rclone/issues/5259