Fix nonblocking read on BSD systems
On FreeBSD and OpenBSD, the syscall.Read call in the anonymous goroutine
in Init was blocking.  This was causing problems during Close: the send
to the quit channel would block, because the anonymous goroutine was
blocked on syscall.Read, and things would remain blocked until
syscall.Read returned.

Init sets the in file descriptor to nonblocking with fcntl, but -- on
FreeBSD and OpenBSD -- the in fd and the out fd are the same, and every
out.Fd call would clear the nonblocking flag.  Fix this problem saving
the out.Fd in a variable prior to setting the in fd to nonblocking, and
use that variable afterward instead of out.Fd.

Related: https://github.com/nsf/termbox-go/issues/167